fix: address review feedback for issue 579
